+++ .github/workflows/CI_endurance.yml
@@ -39,7 +39,7 @@ jobs:
           make libwaku
 
       - name: Build waku bindings
-        run: make -C waku build
+        run: make -C pkg/kernel build
 
       - name: Increase ulimit
         run: sudo sh -c "ulimit -n 8192"
@@ -49,7 +49,7 @@ jobs:
         run: |
           set -euo pipefail
           go test -tags stress  \
-          -p=1 -v ./waku -count=1 -timeout=360m -run '^(TestStressMemoryUsageForThreeNodes|TestStressStoreQuery5kMessagesWithPagination|TestStressHighThroughput10kPublish|TestStressLargePayloadEphemeralMessagesEndurance|TestStressConnectDisconnect1kIteration)$' | tee testlogs1.log
+          -p=1 -v ./pkg/kernel -count=1 -timeout=360m -run '^(TestStressMemoryUsageForThreeNodes|TestStressStoreQuery5kMessagesWithPagination|TestStressHighThroughput10kPublish|TestStressLargePayloadEphemeralMessagesEndurance|TestStressConnectDisconnect1kIteration)$' | tee testlogs1.log
 
       - name: Upload Test Logs (Group 1)
         uses: actions/upload-artifact@v4
@@ -98,7 +98,7 @@ jobs:
           make libwaku
 
       - name: Build waku bindings
-        run: make -C waku build
+        run: make -C pkg/kernel build
 
       - name: Increase ulimit
         run: sudo sh -c "ulimit -n 8192"
@@ -109,7 +109,7 @@ jobs:
         run: |
           set -euo pipefail
           go test -tags stress \
-          -p=1 -v ./waku -count=1 -timeout=360m -run '^(TestStressRandomNodesInMesh|TestStress2Nodes2kIterationTearDown|TestPeerExchangePXLoad)$' | tee testlogs2.log
+          -p=1 -v ./pkg/kernel -count=1 -timeout=360m -run '^(TestStressRandomNodesInMesh|TestStress2Nodes2kIterationTearDown|TestPeerExchangePXLoad)$' | tee testlogs2.log
 
       - name: Upload Test Logs (Group 2)
         uses: actions/upload-artifact@v4
