+++ examples/README.md
@@ -1,0 +1,6 @@
+# Examples
+
+Runnable examples for the Go bindings.
+
+A `messaging/` example (create a node, subscribe, send, print received events)
+will be added alongside the Messaging API implementation.
