+++ pkg/kernel/common/message_hash.go
