+++ pkg/kernel/logging.go
@@ -1,4 +1,4 @@
-package waku
+package kernel
 
 import (
 	"sync"
