+++ pkg/kernel/nodes_basic_test.go
@@ -1,10 +1,10 @@
-package waku
+package kernel
 
 import (
 	"testing"
 	"time"
 
-	"github.com/logos-messaging/logos-delivery-go-bindings/waku/common"
+	"github.com/logos-messaging/logos-delivery-go-bindings/pkg/kernel/common"
 	"github.com/stretchr/testify/require"
 )
