+++ pkg/kernel/nwaku_test_utils.go
@@ -1,4 +1,4 @@
-package waku
+package kernel
 
 import (
 	"context"
@@ -16,9 +16,9 @@ import (
 
 	"github.com/cenkalti/backoff/v3"
 	"github.com/libp2p/go-libp2p/core/peer"
+	"github.com/logos-messaging/logos-delivery-go-bindings/pkg/kernel/common"
+	"github.com/logos-messaging/logos-delivery-go-bindings/pkg/kernel/pb"
 	"github.com/logos-messaging/logos-delivery-go-bindings/utils"
-	"github.com/logos-messaging/logos-delivery-go-bindings/waku/common"
-	"github.com/logos-messaging/logos-delivery-go-bindings/waku/pb"
 	"google.golang.org/protobuf/proto"
 )
