+++ pkg/kernel/store_test.go
@@ -1,4 +1,4 @@
-package waku
+package kernel
 
 import (
 	"fmt"
@@ -6,10 +6,10 @@ import (
 	"time"
 
 	"github.com/cenkalti/backoff/v3"
-	"github.com/logos-messaging/logos-delivery-go-bindings/waku/pb"
+	"github.com/logos-messaging/logos-delivery-go-bindings/pkg/kernel/pb"
 	"github.com/stretchr/testify/require"
 
-	"github.com/logos-messaging/logos-delivery-go-bindings/waku/common"
+	"github.com/logos-messaging/logos-delivery-go-bindings/pkg/kernel/common"
 	"google.golang.org/protobuf/proto"
 )
